+++ src/main/scala/uk/co/developmentanddinosaurs/gatling/lambda/protocol/LambdaProtocolBuilder.scala
@@ -1,4 +1,4 @@
-package uk.co.developmentanddinosaurs.gatlinglambdaextension.protocol
+package uk.co.developmentanddinosaurs.gatling.lambda.protocol
 
 import software.amazon.awssdk.services.lambda.LambdaClient
