+++ src/main/scala/uk/co/developmentanddinosaurs/gatling/lambda/protocol/LambdaProtocolKey.scala
@@ -1,4 +1,4 @@
-package uk.co.developmentanddinosaurs.gatlinglambdaextension.protocol
+package uk.co.developmentanddinosaurs.gatling.lambda.protocol
 
 import io.gatling.core.CoreComponents
 import io.gatling.core.config.GatlingConfiguration
